regulating test folders
