+++ app/waf/test/fixtures/controllers/mocks/mock.base.controller.ts
@@ -1,0 +1,3 @@
+export class MockBaseController {
+  constructor(...args: object[]) {}
+}
